pyproject.toml
[tool.poetry]
name = "imageliner"
version = "0.1.0"
description = ""
authors = ["IoriGunji-Vantiq <igunji@vantiq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^9.5.0"
webcolors = "^1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
